Pizza/requirement.txt
aniso8601==9.0.1
attrs==22.2.0
autopep8==2.0.1
click==8.1.3
Flask==2.2.2
flask-restx==1.0.5
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
pycodestyle==2.10.0
pyrsistent==0.19.3
python-decouple==3.7
python-dotenv==0.21.1
pytz==2022.7.1
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.12.0
